JE2-38 refactoring: modules, controllers, and services are reorganized into respective folders
