+++ api/src/modules/clients.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common'
+import { ClientsController } from 'src/controllers/clients.controller'
+import { DbModule } from './db.module'
+
+@Module({
+  imports: [DbModule],
+  controllers: [ClientsController],
+})
+export class ClientsModule {}
